Add files via upload
The SU1 application is an all-in-one mobile solution designed to enhance the experience of international travelers. It features a chatbot assistance to help the user in traveling situations and making a travel planner. It is also an intelligent translator that takes cultural norms into account, a dynamic planner that adjusts to weather and schedules, real-time cost analysis, and the automatic creation of personalized itineraries. The goal is to break down linguistic and cultural barriers, optimize the budget, and simplify trip planning.